Trimmed project name down to iodaconv. Put WriteNcVar routine in the lib-python module.
