+++ src/CMakeLists.txt
@@ -1,2 +1,3 @@
+add_subdirectory(lib-python)
 add_subdirectory(gsi-ncdiag)
 add_subdirectory(ncep)
