+++ src/gsi-ncdiag/prep_nc_files.py
@@ -8,6 +8,7 @@ import argparse
 import numpy as np
 import netCDF4
 from netCDF4 import Dataset
+import ioda_conv_ncio
 
 ###################################################################################
 # SUBROUTINES
@@ -29,21 +30,6 @@ def ReshapeVar(Var, Nlocs, Nchans):
         OutDims = ('nlocs', 'nchans', Var.dimensions[1])
 
     return [ OutVals, OutDims ]
-
-def WriteNcVar(Gfid, Ofid, OutDest, Vname, Vdtype, Vdims, Vvals):
-    ###############################################################
-    # This method will write out the variable into the appropriate
-    # netcdf files.
-    #
-    if (OutDest == 'G' or OutDest == 'B'):
-        # write to geo file
-        Ovar = Gfid.createVariable(Vname, Vdtype, Vdims)
-        Ovar[...] = Vvals[...]
-
-    if (OutDest == 'O' or OutDest == 'B'):
-        # write to obs file
-        Ovar = Ofid.createVariable(Vname, Vdtype, Vdims)
-        Ovar[...] = Vvals[...]
 
 ###################################################################################
 # CLASSES
