+++ src/lib-python/CMakeLists.txt
@@ -1,0 +1,3 @@
+install (PROGRAMS
+  ioda_conv_ncio.py
+  DESTINATION bin)
